requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
nfcpy==1.0.4
ndef==0.3.3
python-dotenv==1.0.0
requests==2.31.0
sqlalchemy==2.0.23